Revert "Ft/employee details dr"
