+++ api/database.json
@@ -4,95 +4,95 @@
         "id": 1,
         "fullName": "John Doe",
         "email": "johndoe@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 2,
         "fullName": "Jane Smith",
         "email": "janesmith@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 3,
         "fullName": "Alice Johnson",
         "email": "alicejohnson@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 4,
         "fullName": "Bob Brown",
         "email": "bobbrown@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 5,
         "fullName": "Carol White",
         "email": "carolwhite@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 6,
         "fullName": "David Lee",
         "email": "davidlee@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 7,
         "fullName": "Ella Davis",
         "email": "elladavis@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 8,
         "fullName": "Frank Harris",
         "email": "frankharris@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 9,
         "fullName": "Grace King",
         "email": "graceking@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 10,
         "fullName": "Henry Moore",
         "email": "henrymoore@example.com",
-        "isAdmin": true
+        "isStaff": true
       },
       {
         "id": 11,
         "fullName": "Isla Scott",
         "email": "islascott@example.com",
-        "isAdmin": false
+        "isStaff": false
       },
       {
         "id": 12,
         "fullName": "Jack Taylor",
         "email": "jacktaylor@example.com",
-        "isAdmin": false
+        "isStaff": false
       },
       {
         "id": 13,
         "fullName": "Katie Wilson",
         "email": "katiewilson@example.com",
-        "isAdmin": false
+        "isStaff": false
       },
       {
         "id": 14,
         "fullName": "Leo Martinez",
         "email": "leomartinez@example.com",
-        "isAdmin": false
+        "isStaff": false
       },
       {
         "id": 15,
         "fullName": "Mia Anderson",
         "email": "miaanderson@example.com",
-        "isAdmin": false
+        "isStaff": false
       }
     ],
 
-    "admins": [
+    "admin": [
         {
           "id": 1,
           "rate": 5,
@@ -172,46 +172,6 @@
           "address": "456 Spruce St, Baytown",
           "phoneNumber": 8887654321,
           "userId": 10
-        },
-        {
-          "id": 11,
-          "rate": 5,
-          "title": "Admin",
-          "address": "789 Oak St, Rivertown",
-          "phoneNumber": 8881234567,
-          "userId": 11
-        },
-        {
-          "id": 12,
-          "rate": 3,
-          "title": "Admin",
-          "address": "321 Birch St, Hillview",
-          "phoneNumber": 8882345678,
-          "userId": 12
-        },
-        {
-          "id": 13,
-          "rate": 4,
-          "title": "Admin",
-          "address": "654 Pine St, Lakeside",
-          "phoneNumber": 8883456789,
-          "userId": 13
-        },
-        {
-          "id": 14,
-          "rate": 2,
-          "title": "Admin",
-          "address": "987 Cedar St, Mountainview",
-          "phoneNumber": 8884567890,
-          "userId": 14
-        },
-        {
-          "id": 15,
-          "rate": 5,
-          "title": "Admin",
-          "address": "159 Maple St, Creekside",
-          "phoneNumber": 8885678901,
-          "userId": 15
         }
       ],
