--- src/components/employees/Employee.css
@@ -1,51 +1,0 @@
-.users {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    margin: 3rem;
-  }
-
-  .users>* {
-    flex-basis: 31%;
-    margin: 0.5rem;
-  }
-
-  section.user {
-    margin: 3rem;
-  }
-
-  .user-details {
-    margin: 3rem 10rem 10rem 10rem;
-  }
-
-  .save-btn {
-    float: right;
-    background-color: var(--primary);
-  }
-
-  .user {
-    background-color: var(--white);
-    border: 1px solid var(--outline);
-    border-radius: 0.5rem;
-    padding: 1rem;
-    box-shadow: 1px 2px 5px lightgray;
-  }
-
-  .user:hover {
-    transform: scale(1.05);
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-    border-color: var(--primary);
-  }
-
-  .user-info {
-    color: var(--info);
-    font-family: "Roboto", sans-serif;
-  }
-
-  .user-header {
-    font-size: larger;
-    font-weight: 700;
-    line-height: 3rem;
-    color: var(--primary);
-  }
