--- src/components/employees/EmployeeDetails.css
@@ -1,22 +1,0 @@
-
-
-  .details {
-    background-color: var(--white);
-    border: 1px solid var(--outline);
-    border-radius: 0.5rem;
-    padding: 1rem;
-    box-shadow: 1px 2px 5px lightgray;
-    display: center;
-  }
-
-  .details-info {
-    color: var(--info);
-    font-family: "Roboto", sans-serif;
-  }
-
-  .details-header {
-    font-size: larger;
-    font-weight: 700;
-    line-height: 3rem;
-    color: var(--primary);
-  }
