--- src/components/users/User.css
@@ -1,17 +1,0 @@
-.user {
-    background-color: var(--white);
-    border: 1px solid var(--outline);
-    border-radius: 0.5rem;
-    padding: 1rem;
-  }
-  
-  .user-info {
-    color: var(--info);
-    font-family: "Roboto", sans-serif;
-  }
-  
-  .user:hover {
-    -webkit-transform: scale(1.1);
-    -ms-transform: scale(1.1);
-    transform: scale(1.1);
-  }
