--- src/components/users/User.jsx
@@ -1,14 +1,0 @@
-import "./User.css"
-
-export const User = ({ user }) => {
-    return (
-        <div className="user">
-            <div>
-                <div className="user-info">{user.fullName}</div>
-            </div>
-            <div>
-                <div className="user-info">{user.email}</div>          
-            </div>
-        </div>
-    )
-}
