+++ src/views/ApplicationViews.jsx
@@ -2,8 +2,6 @@ import { useEffect, useState } from "react"
 import { Outlet, Route, Routes } from "react-router-dom"
 import { NavBar } from "../components/navbar/NavBar"
 import { Welcome } from "../components/welcome/Welcome"
-import { EmployeeList } from "../components/employees/EmployeeList"
-import { EmployeeDetails } from "../components/employees/EmployeeDetails"
 
 
 
@@ -22,8 +20,8 @@ export const ApplicationViews = () => {
       }>
         <Route index element={<Welcome />} />
           <Route path="employees">
-          <Route index element={<EmployeeList />} />
-          <Route path=":employeeId" element={<EmployeeDetails/>} />
+          <Route index element={<>Employee List</>} />
+          <Route path=":employeeId" element={<>Employee Details</>} />
           <Route path="edit/:employeeId" element={<>Edit Employee Details</>} />
 
         </Route>    
@@ -32,4 +30,3 @@ export const ApplicationViews = () => {
   </>)
 
 }
-
